Add files via upload
This configuration suits Delta Wasp 2040 Turbo 2. The hotend temperature is not working, mainly due to the wiring, which I haven't completely figured out yet. The stock AD595 may not be ideally suited to a common positive with NPN switching on the heater. There's a common line here for the temp sensor chip.

I had tuned the radius to my machine. You should use delta auto config with a probe to calibrate it to your machine. My setup uses TMC2209 drivers and a RepRap discount LCD. Changes like these ones will need to be adapted to your setup. This is a work in progress. To use this, I recommend using a multimeter to check the wiring and find out where things all go to. That's basically how I got this working.